+++ server/plugin.go
@@ -2,14 +2,12 @@ package main
 
 import (
 	"embed"
-	"fmt"
+
 	"net/http"
 	"os"
 	"os/exec"
 	"sync"
 	"time"
-
-	"errors"
 
 	sq "github.com/Masterminds/squirrel"
 	"github.com/jmoiron/sqlx"
@@ -19,7 +17,6 @@ import (
 	"github.com/mattermost/mattermost-plugin-ai/server/llm"
 	"github.com/mattermost/mattermost-plugin-ai/server/metrics"
 	"github.com/mattermost/mattermost-plugin-ai/server/openai"
-	"github.com/mattermost/mattermost/server/public/model"
 	"github.com/mattermost/mattermost/server/public/plugin"
 	"github.com/mattermost/mattermost/server/public/pluginapi"
 	"github.com/mattermost/mattermost/server/public/shared/httpservice"
@@ -179,106 +176,5 @@ func (p *Plugin) getTranscribe() Transcriber {
 	case "azure":
 		return openai.NewAzure(botConfig.Service, p.llmUpstreamHTTPClient, llmMetrics)
 	}
-	return nil
-}
-
-var (
-	// ErrNoResponse is returned when no response is posted under a normal condition.
-	ErrNoResponse = errors.New("no response")
-)
-
-func (p *Plugin) MessageHasBeenPosted(c *plugin.Context, post *model.Post) {
-	if err := p.handleMessages(post); err != nil {
-		if errors.Is(err, ErrNoResponse) {
-			p.pluginAPI.Log.Debug(err.Error())
-		} else {
-			p.pluginAPI.Log.Error(err.Error())
-		}
-	}
-}
-
-const (
-	ActivateAIProp  = "activate_ai"
-	FromWebhookProp = "from_webhook"
-	FromBotProp     = "from_bot"
-	FromPluginProp  = "from_plugin"
-	WranglerProp    = "wrangler"
-)
-
-func (p *Plugin) handleMessages(post *model.Post) error {
-	// Don't respond to ourselves
-	if p.IsAnyBot(post.UserId) {
-		return fmt.Errorf("not responding to ourselves: %w", ErrNoResponse)
-	}
-
-	// Never respond to remote posts
-	if post.RemoteId != nil && *post.RemoteId != "" {
-		return fmt.Errorf("not responding to remote posts: %w", ErrNoResponse)
-	}
-
-	// Wrangler posts should be ignored
-	if post.GetProp(WranglerProp) != nil {
-		return fmt.Errorf("not responding to wrangler posts: %w", ErrNoResponse)
-	}
-
-	// Don't respond to plugins unless they ask for it
-	if post.GetProp(FromPluginProp) != nil && post.GetProp(ActivateAIProp) == nil {
-		return fmt.Errorf("not responding to plugin posts: %w", ErrNoResponse)
-	}
-
-	// Don't respond to webhooks
-	if post.GetProp(FromWebhookProp) != nil {
-		return fmt.Errorf("not responding to webhook posts: %w", ErrNoResponse)
-	}
-
-	channel, err := p.pluginAPI.Channel.Get(post.ChannelId)
-	if err != nil {
-		return fmt.Errorf("unable to get channel: %w", err)
-	}
-
-	postingUser, err := p.pluginAPI.User.Get(post.UserId)
-	if err != nil {
-		return err
-	}
-
-	// Don't respond to other bots unless they ask for it
-	if (postingUser.IsBot || post.GetProp(FromBotProp) != nil) && post.GetProp(ActivateAIProp) == nil {
-		return fmt.Errorf("not responding to other bots: %w", ErrNoResponse)
-	}
-
-	// Check we are mentioned like @ai
-	if bot := p.GetBotMentioned(post.Message); bot != nil {
-		return p.handleMentions(bot, post, postingUser, channel)
-	}
-
-	// Check if this is post in the DM channel with any bot
-	if bot := p.GetBotForDMChannel(channel); bot != nil {
-		return p.handleDMs(bot, channel, postingUser, post)
-	}
-
-	return nil
-}
-
-func (p *Plugin) handleMentions(bot *Bot, post *model.Post, postingUser *model.User, channel *model.Channel) error {
-	if err := p.checkUsageRestrictions(postingUser.Id, bot, channel); err != nil {
-		return err
-	}
-
-	if err := p.processUserRequestToBot(bot, p.MakeConversationContext(bot, postingUser, channel, post)); err != nil {
-		return fmt.Errorf("unable to process bot mention: %w", err)
-	}
-
-	return nil
-}
-
-func (p *Plugin) handleDMs(bot *Bot, channel *model.Channel, postingUser *model.User, post *model.Post) error {
-	if err := p.checkUsageRestrictionsForUser(bot, postingUser.Id); err != nil {
-		return err
-	}
-
-	if err := p.processUserRequestToBot(bot, p.MakeConversationContext(bot, postingUser, channel, post)); err != nil {
-		return fmt.Errorf("unable to process bot DM: %w", err)
-	}
-
 	return nil
 }
